remake sidebar
