+++ src/App.js
@@ -1,10 +1,14 @@
 import "./app.css";
-import Search from "./pages/Search";
+// import Search from "./pages/Search";
+// import CreatePlaylist from "./pages/CreatePlaylist";
+import Sidebar from "./components/sidebar";
 
 function App() {
   return (
     <div className="app">
-      <Search />
+      <Sidebar />
+      {/* <Search /> */}
+      {/* <CreatePlaylist /> */}
     </div>
   );
 }
