+++ src/components/sidebar/index.js
@@ -3,8 +3,14 @@ import style from "./sidebar.module.css";
 
 export default function Sidebar() {
   return (
-    <div className={style.sidebar}>
-      <button> + Create Playlist</button>
+    <div className={style["wrapper-sidebar"]}>
+      <div className={style["logo"]}>
+        <h2>Music Player</h2>
+      </div>
+      <div className={style["menu"]}>
+        <button> Create Playlist</button>
+        <button> Search</button>
+      </div>
     </div>
   );
 }
