+++ src/components/sidebar/sidebar.module.css
@@ -1,27 +1,32 @@
-.sidebar {
-  /* background-color: #222831; */
-  background-color: black;
+.wrapper-sidebar {
+  background-color: #000000;
   color: whitesmoke;
   width: 280px;
-  height: auto;
-  padding: 30px;
+  min-height: 100vh;
+  display: grid;
+  grid-template-rows: 70px 1fr;
+  row-gap: 20px;
 }
 
-.sidebar > button {
-  letter-spacing: 1px;
-  border-style: none;
-  padding: 10px;
-  width: 100%;
-  border-radius: 6px;
-  font-size: 14px;
-  font-weight: bold;
+.logo {
+  height: 70px;
+}
+
+.logo h2 {
+  text-align: center;
+  line-height: 70px;
+}
+
+.menu button {
+  background-color: #000000;
+  min-width: 100%;
+  min-height: 50px;
   cursor: pointer;
-  background-color: black;
-  color: whitesmoke;
-  text-align: left;
-  margin-bottom: 10px;
+  color: gainsboro;
+  border: none;
+  font-weight: bold;
 }
 
-.sidebar > button:hover {
-  background-color: #222831;
+.menu button:hover {
+  color: white;
 }
