--- src/components/track/track.module.css
@@ -1,17 +1,0 @@
-.track_list {
-  margin: 30px;
-  padding: 10px;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.list-track {
-  background-color: #121212;
-  color: #fff;
-  width: 80vw;
-}
-
-.list-track thead {
-  background-color: black;
-  padding: 200px;
-}
